rename `merge` to `ext`
Renamed `rag merge` command to `rag ext`.

`rag merge` was misleading that `git merge` merges another branch to a
branch, while `rag merge` merges a knowledge-base (a repo for a git). To
be more like git, it has to be `rag submodule`, but the term `module` is
still misleading. I chose `ext` because it "extends" the current
knowledge-base with an "external" knowledge-base.