change behavior of `rag build`
While fixing `tests/svg.py`, I decided to change the behavior of `rag
build`. Previously, it returned immediately if any of files fails to
build. The remaining files (likely to be okay) are kept staged.

Now, it skips failed files and continues to process the other files. It
always iterates all the files in `index.staged_files` and returns all
the errors in the end.

Now that there's no need for "atomic build", it doesn't have to scan the
files before it builds a knowledge-base. So I removed the unnecessary
loop.