add error handling to `tests/tests.py`
There's an edge case was clean up process of the test harness. When a
test dies, it removes tmp directories and changes directory to root. But
sometimes, a test case spawns children and dies while the children are
still running. The children are still writing something to tmp
directories and it would mess up `shutil.rmtree()`. So I added another
`try-catch` to the cleanup code, and added the error message to the
result.

The error messages from cleanup code are not visible from the ci
dashboard, but we can still see them when we open the result file.