kill `server_process` more smartly
Tests that spawns `server_process` are supposed to kill the process
before it exits. Otherwise, other tests that also need server will not
run.

Previously, it naively did `finally: server_process.kill()`. It worked
because any exceptions will be caught later. But it hides the *real*
exception and throws "server_process is not assigned". The tester has no
idea why the test failed.