fix `tests/images.py`
The test adds 3 images to a knowledge-base and checks whether ragit can
process all the 3 images correctly. It failed because there're only 2
images left in the knowledge-base. It's because the images are
normalized, and an empty images becomes the same after the
normalization. (I'm still not sure why it sometimes succeeds and
sometimes fails).

I fixed it by replacing empty images with different images.