fix pdl escape rule
Previously, pdl's escape rule was very unintuitive. You had to turn on
or off escapes each time, and it's very tricky to use pdl tokens in
special ways.

Now it works like how html escape works with web browsers and template
engines. Escaping is always on, and all you have to do is to use `&lt;`
and `safe` filter.