refactor pdl escape rules, again
Previously, `RenderableChunk` had very ambiguous semantics and name.
Some of its contents are escaped and the others are not. In order to use
this type, you had to postprocess its contents. If you forget to do so,
it'll cause a very hard-to-catch bug.

Now, it explicitly separates escaped contents and unescaped contents.
You can choose one that you need.