rename `renderable.rs` to `render.rs`
Renamed `src/chunk/renderable.rs` to `src/chunk/render.rs`. It'd be nice
to squash it with the previous commit, but I didn't. It's because it's a
good practice to rename a file in a commit and change its contents in
another commit, so that git can detect file renames easier.